Create separate controllers for mobile commons interatction and CRUD admin interface
