+++ app/Http/Controllers/MessageController.php
@@ -7,78 +7,78 @@
 
   class MessageController extends Controller
   {
+    /**
+     * Display a listing of the resource.
+     *
+     * @return Response
+     */
     public function index()
     {
-      // Get the mdata the users is in.
-      $mdata_id = $this->mdata_id;
-
-      if ($this->is_mms_msg) {
-        if ($mdata_id == '12368') {
-          $matched_response = '195188';
-        }
-        elseif ($mdata_id == '12388') {
-          $matched_response = '195456';
-        }
-
-        return $this->mobile_commons->sendMessage($matched_response, $this->phone);
-      }
-
-      // Get the user message.
-      $user_message = $this->args;
-      $user_message = Message::sanitizeMessage($user_message);
-
-      // Test phrase messages.
-      if ($matched_response = Message::testPhraseMessages($user_message, $mdata_id)) {
-        $matched_response = self::getFinalOptInPath($matched_response);
-      }
-      // Test word messages.
-      else if ($matched_response = Message::testWordMessages($user_message, $mdata_id)) {
-        $matched_response = self::getFinalOptInPath($matched_response);
-      }
-      // Test regex messages.
-      else if ($matched_response = Message::testRegexMessages($user_message, $mdata_id)) {
-        $matched_response = self::getFinalOptInPath($matched_response);
-      }
-      // No match found message.
-      else {
-        $matched_response = Message::determineOptInPath('default', $mdata_id);
-      }
-
-      return $this->mobile_commons->sendMessage($matched_response, $this->phone);
+        //
     }
 
-    /*
-     * If there are multiple response options, this function will
-     * select a random one to use.
+    /**
+     * Show the form for creating a new resource.
      *
-     * @param string $matched_response - A string of opt-in path IDs.
+     * @return Response
      */
-    public function getFinalOptInPath($matched_response) {
-      $options = explode(',', $matched_response);
-      $count = count($options);
-
-      if ($count > 1) {
-        $index = array_rand($options);
-        return trim($options[$index]);
-      }
-
-      return $matched_response;
+    public function create()
+    {
+        //
     }
 
-    /*
-    * Grabs all of the opt in paths for the pregnancy text campaign
-    * via the mobile commons api.
-    *
-    * @return JSON
-    */
-    public function getPaths() {
-      $paths = $this->mobile_commons->getCampaignOptInPaths($this->campaign_id);
-      $path_ids = array();
-      foreach ($paths as $key => $message) {
-        $message = strtolower($message);
-        $path_ids[$key] = $message;
-      }
-
-      return response()->json($path_ids);
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @return Response
+     */
+    public function store()
+    {
+        //
     }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return Response
+     */
+    public function show($id)
+    {
+        //
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return Response
+     */
+    public function edit($id)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  int  $id
+     * @return Response
+     */
+    public function update($id)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return Response
+     */
+    public function destroy($id)
+    {
+        //
+    }
+
   }
